Move logic for refreshing tokens to auth context
